Update StudentControllerTest and StudentIT
